use size_t/ptrdiff_t for total #points

+++ src/bicgstab.hpp
@@ -24,13 +24,13 @@ namespace meep {
 
 typedef void (*bicgstab_op)(const realnum *x, realnum *y, void *data);
 
-int bicgstabL(const int L, 
-	      const int n, realnum *x,
-	      bicgstab_op A, void *Adata, const realnum *b,
-	      const double tol, 
-	      int *iters, // input *iters = max iters, output = actual iters
-	      realnum *work, // if you pass work=NULL, bicgstab returns nwork
-	      const bool quiet);
+ptrdiff_t bicgstabL(const int L,
+              const size_t n, realnum *x,
+              bicgstab_op A, void *Adata, const realnum *b,
+              const double tol,
+              int *iters, // input *iters = max iters, output = actual iters
+              realnum *work, // if you pass work=NULL, bicgstab returns nwork
+              const bool quiet);
 
 } // namespace meep
