+++ src/multilevel-atom.cpp
@@ -79,7 +79,7 @@ multilevel_susceptibility::~multilevel_susceptibility() {
 #if MEEP_SINGLE
 #  define DGETRF F77_FUNC(sgetrf,SGETRF)
 #  define DGETRI F77_FUNC(sgetri,SGETRI)
-#else 
+#else
 #  define DGETRF F77_FUNC(dgetrf,DGETRF)
 #  define DGETRI F77_FUNC(dgetri,DGETRI)
 #endif
@@ -95,7 +95,7 @@ static bool invert(realnum *S, int p)
   DGETRF(&p, &p, S, &p, ipiv, &info);
   if (info < 0) abort("invalid argument %d in DGETRF", -info);
   if (info > 0) { delete[] ipiv; return false; } // singular
-  
+
   int lwork = -1;
   realnum work1;
   DGETRI(&p, S, &p, ipiv, &work1, &lwork, &info);
@@ -117,7 +117,7 @@ static bool invert(realnum *S, int p)
 typedef realnum *realnumP;
 typedef struct {
   size_t sz_data;
-  int ntot;
+  size_t ntot;
   realnum *GammaInv; // inv(1 + Gamma * dt / 2)
   realnumP *P[NUM_FIELD_COMPONENTS][2]; // P[c][cmp][transition][i]
   realnumP *P_prev[NUM_FIELD_COMPONENTS][2];
@@ -129,7 +129,7 @@ typedef struct {
 void *multilevel_susceptibility::new_internal_data(
 				    realnum *W[NUM_FIELD_COMPONENTS][2],
 				    const grid_volume &gv) const {
-  int num = 0; // number of P components
+  size_t num = 0; // number of P components
   FOR_COMPONENTS(c) DOCMP2 if (needs_P(c, cmp, W)) num += 2 * gv.ntot();
   size_t sz = sizeof(multilevel_data)
     + sizeof(realnum) * (L*L + L + gv.ntot()*L + num*T - 1);
@@ -146,7 +146,7 @@ void multilevel_susceptibility::init_internal_data(
   size_t sz_data = d->sz_data;
   memset(d, 0, sz_data);
   d->sz_data = sz_data;
-  int ntot = d->ntot = gv.ntot();
+  size_t ntot = d->ntot = gv.ntot();
 
   /* d->data points to a big block of data that holds GammaInv, P,
      P_prev, Ntmp, and N.  We also initialize a bunch of convenience
@@ -158,7 +158,7 @@ void multilevel_susceptibility::init_internal_data(
   for (int i = 0; i < L; ++i)
     for (int j = 0; j < L; ++j)
       d->GammaInv[i*L + j] = (i == j) + Gamma[i*L + j] * dt/2;
-  if (!invert(d->GammaInv, L)) 
+  if (!invert(d->GammaInv, L))
     abort("multilevel_susceptibility: I + Gamma*dt/2 matrix singular");
 
   realnum *P = d->data + L*L;
@@ -178,7 +178,7 @@ void multilevel_susceptibility::init_internal_data(
   d->N = P + L; // the last L*ntot block of the data
 
   // initial populations
-  for (int i = 0; i < ntot; ++i)
+  for (size_t i = 0; i < ntot; ++i)
     for (int l = 0; l < L; ++l)
       d->N[i*L + l] = N0[l];
 }
@@ -199,7 +199,7 @@ void *multilevel_susceptibility::copy_internal_data(void *data) const {
   if (!d) return 0;
   multilevel_data *dnew = (multilevel_data *) malloc(d->sz_data);
   memcpy(dnew, d, d->sz_data);
-  int ntot = d->ntot;
+  size_t ntot = d->ntot;
   dnew->GammaInv = dnew->data;
   realnum *P = dnew->data + L*L;
   realnum *P_prev = P + ntot;
@@ -225,8 +225,8 @@ int multilevel_susceptibility::num_cinternal_notowned_needed(component c,
 }
 
 realnum *multilevel_susceptibility::cinternal_notowned_ptr(
-				        int inotowned, component c, int cmp, 
-					int n, 
+				        int inotowned, component c, int cmp,
+					int n,
 					void *P_internal_data) const {
   multilevel_data *d = (multilevel_data *) P_internal_data;
   if (!d->P[c][cmp] || inotowned < 0 || inotowned >= T) // never true
@@ -236,14 +236,14 @@ realnum *multilevel_susceptibility::cinternal_notowned_ptr(
 
 void multilevel_susceptibility::update_P
        (realnum *W[NUM_FIELD_COMPONENTS][2],
-	realnum *W_prev[NUM_FIELD_COMPONENTS][2], 
+	realnum *W_prev[NUM_FIELD_COMPONENTS][2],
 	double dt, const grid_volume &gv, void *P_internal_data) const {
   multilevel_data *d = (multilevel_data *) P_internal_data;
   double dt2 = 0.5 * dt;
 
   // field directions and offsets for E * dP dot product.
   component cdot[3] = {Dielectric,Dielectric,Dielectric};
-  int o1[3], o2[3];
+  ptrdiff_t o1[3], o2[3];
   int idot = 0;
   FOR_COMPONENTS(c) if (d->P[c][0]) {
     if (idot == 3) abort("bug in meep: too many polarization components");
@@ -256,7 +256,7 @@ void multilevel_susceptibility::update_P
   realnum *Ntmp = d->Ntmp;
   LOOP_OVER_VOL_OWNED(gv, Centered, i) {
     realnum *N = d->N + i*L; // N at current point, to update
-    
+
     // Ntmp = (I - Gamma * dt/2) * N
     for (int l1 = 0; l1 < L; ++l1) {
       Ntmp[l1] = (1.0 - Gamma[l1*L + l1]*dt2) * N[l1]; // diagonal term
@@ -326,11 +326,11 @@ void multilevel_susceptibility::update_P
       if (w && s) {
 	realnum *p = d->P[c][cmp][t], *pp = d->P_prev[c][cmp][t];
 
-	int o1, o2;
+	ptrdiff_t o1, o2;
 	gv.cent2yee_offsets(c, o1, o2);
 	o1 *= L; o2 *= L;
 	const realnum *N = d->N;
-	
+
 	// directions/strides for offdiagonal terms, similar to update_eh
 	const direction d = component_direction(c);
 	direction d1 = cycle_direction(gv.dim, d, 1);
@@ -341,7 +341,7 @@ void multilevel_susceptibility::update_P
 	component c2 = direction_component(c, d2);
 	const realnum *w2 = W[c2][cmp];
 	const realnum *s2 = w2 ? sigma[c][d2] : NULL;
-	
+
 	if (s1 || s2) {
 	  abort("nondiagonal saturable gain is not yet supported");
 	}
@@ -352,8 +352,8 @@ void multilevel_susceptibility::update_P
 	    // dNi is population inversion for this transition
 	    double dNi = -0.25 * (Ni[lp]+Ni[lp+o1]+Ni[lp+o2]+Ni[lp+o1+o2]
 				  -Ni[lm]-Ni[lm+o1]-Ni[lm+o2]-Ni[lm+o1+o2]);
-	    p[i] = gamma1inv * (pcur * (2 - omega0dtsqr) 
-				- gamma1 * pp[i] 
+	    p[i] = gamma1inv * (pcur * (2 - omega0dtsqr)
+				- gamma1 * pp[i]
 				+ omega0dtsqr * (st * s[i] * w[i])) * dNi;
 	    pp[i] = pcur;
 	  }
@@ -364,18 +364,18 @@ void multilevel_susceptibility::update_P
 }
 
 void multilevel_susceptibility::subtract_P(field_type ft,
-			  realnum *f_minus_p[NUM_FIELD_COMPONENTS][2], 
+			  realnum *f_minus_p[NUM_FIELD_COMPONENTS][2],
 					   void *P_internal_data) const {
   multilevel_data *d = (multilevel_data *) P_internal_data;
   field_type ft2 = ft == E_stuff ? D_stuff : B_stuff; // for sources etc.
-  int ntot = d->ntot;
-  for (int t = 0; t < T; ++t) { 
+  size_t ntot = d->ntot;
+  for (int t = 0; t < T; ++t) {
     FOR_FT_COMPONENTS(ft, ec) DOCMP2 if (d->P[ec][cmp]) {
       component dc = field_type_component(ft2, ec);
       if (f_minus_p[dc][cmp]) {
 	realnum *p = d->P[ec][cmp][t];
 	realnum *fmp = f_minus_p[dc][cmp];
-	for (int i = 0; i < ntot; ++i) fmp[i] -= p[i];
+	for (size_t i = 0; i < ntot; ++i) fmp[i] -= p[i];
       }
     }
   }
